refactors nav by removing projectsNav
