+++ client/components/Header.tsx
@@ -1,8 +1,11 @@
 function Header() {
   return (
-    <div>
-      <h1>MAIA TARAPIPIPI</h1>
-      <h2>Full-Stack Software Developer | Textile Artist</h2>
+    <div className="">
+      <div className="flex justify-center text-4xl">MAIA TARAPIPIPI</div>
+
+      <div className="flex justify-center text-2xl">
+        Full-Stack Software Developer | Textile Artist
+      </div>
     </div>
   )
 }
